refactor: create FedimintApi for common Api functions
