+++ fedimint-server/src/consensus/interconnect.rs
@@ -4,7 +4,8 @@ use fedimint_core::module::interconnect::ModuleInterconect;
 use fedimint_core::module::{ApiError, ApiRequestErased};
 use serde_json::Value;
 
-use crate::net::api::{ConsensusApi, HasApiContext};
+use crate::api::HasApiContext;
+use crate::net::api::ConsensusApi;
 
 pub struct FedimintInterconnect<'a> {
     pub fedimint: &'a ConsensusApi,
