+++ fedimint-server/src/net/api.rs
@@ -1,10 +1,7 @@
 //! Implements the client API through which users interact with the federation
 use std::fmt::{Debug, Formatter};
-use std::panic::AssertUnwindSafe;
 use std::sync::Arc;
-use std::time::Duration;
 
-use anyhow::Context;
 use async_trait::async_trait;
 use bitcoin_hashes::sha256;
 use fedimint_core::api::WsClientConnectInfo;
@@ -18,7 +15,6 @@ use fedimint_core::module::{
 };
 use fedimint_core::outcome::TransactionStatus;
 use fedimint_core::server::DynServerModule;
-use fedimint_core::task::TaskHandle;
 use fedimint_core::transaction::Transaction;
 use fedimint_core::{OutPoint, TransactionId};
 use fedimint_logging::{LOG_NET_API, LOG_TASK};
@@ -29,8 +25,9 @@ use jsonrpsee::types::ErrorObject;
 use jsonrpsee::RpcModule;
 use tokio::sync::mpsc::error::SendError;
 use tokio::sync::mpsc::Sender;
-use tracing::{debug, error};
+use tracing::debug;
 
+use crate::api::HasApiContext;
 use crate::config::api::ApiResult;
 use crate::config::ServerConfig;
 use crate::consensus::interconnect::FedimintInterconnect;
@@ -280,19 +277,6 @@ impl ConsensusApi {
     }
 }
 
-/// Has the context necessary for serving API endpoints
-///
-/// Returns the specific `State` the endpoint requires and the
-/// `ApiEndpointContext` which all endpoints can access.
-#[async_trait]
-pub trait HasApiContext<State> {
-    async fn context(
-        &self,
-        request: &ApiRequestErased,
-        id: Option<ModuleInstanceId>,
-    ) -> (&State, ApiEndpointContext<'_>);
-}
-
 #[async_trait]
 impl HasApiContext<ConsensusApi> for ConsensusApi {
     async fn context(
@@ -332,116 +316,7 @@ impl HasApiContext<DynServerModule> for ConsensusApi {
     }
 }
 
-pub async fn run_server(cfg: ServerConfig, fedimint: Arc<ConsensusApi>, task_handle: TaskHandle) {
-    let state = RpcHandlerCtx {
-        rpc_context: fedimint.clone(),
-    };
-    let mut rpc_module = RpcModule::new(state);
-
-    attach_endpoints(&mut rpc_module, server_endpoints(), None);
-
-    for (id, module) in fedimint.modules.iter_modules() {
-        attach_endpoints(&mut rpc_module, module.api_endpoints(), Some(id));
-    }
-
-    debug!(addr = cfg.local.api_bind.to_string(), "Starting WSServer");
-    let server = ServerBuilder::new()
-        .max_connections(cfg.local.max_connections)
-        .ping_interval(Duration::from_secs(10))
-        .build(&cfg.local.api_bind.to_string())
-        .await
-        .context(format!("Bind address: {}", cfg.local.api_bind))
-        .expect("Could not start API server");
-
-    let server_handle = server
-        .start(rpc_module)
-        .expect("Could not start API server");
-
-    let stop_handle = server_handle.clone();
-
-    task_handle
-        .on_shutdown(Box::new(move || {
-            Box::pin(async move {
-                debug!(target: LOG_TASK, "Shutting down jsonrpcsee server");
-                // ignore errors: we don't care if already stopped
-                let _ = stop_handle.stop();
-            })
-        }))
-        .await;
-
-    server_handle.stopped().await;
-    debug!(target: LOG_TASK, "jsonrpcsee server stopped");
-}
-
-const API_ENDPOINT_TIMEOUT: Duration = Duration::from_secs(60);
-
-/// Attaches `endpoints` to the `RpcModule`
-pub fn attach_endpoints<State, T>(
-    rpc_module: &mut RpcModule<RpcHandlerCtx<T>>,
-    endpoints: Vec<ApiEndpoint<State>>,
-    module_instance_id: Option<ModuleInstanceId>,
-) where
-    T: HasApiContext<State> + Sync + Send + 'static,
-    State: Sync + Send + 'static,
-{
-    for endpoint in endpoints {
-        let path = if let Some(module_instance_id) = module_instance_id {
-            // This memory leak is fine because it only happens on server startup
-            // and path has to live till the end of program anyways.
-            Box::leak(format!("/module/{}{}", module_instance_id, endpoint.path).into_boxed_str())
-        } else {
-            endpoint.path
-        };
-
-        // Another memory leak that is fine because the function is only called once at
-        // startup
-        let handler: &'static _ = Box::leak(endpoint.handler);
-
-        rpc_module
-            .register_async_method(path, move |params, rpc_state| async move {
-                let params = params.one::<serde_json::Value>()?;
-                let rpc_context = &rpc_state.rpc_context;
-
-                // Using AssertUnwindSafe here is far from ideal. In theory this means we could
-                // end up with an inconsistent state in theory. In practice most API functions
-                // are only reading and the few that do write anything are atomic. Lastly, this
-                // is only the last line of defense
-                AssertUnwindSafe(tokio::time::timeout(API_ENDPOINT_TIMEOUT, async {
-                    let request = serde_json::from_value(params)
-                        .map_err(|e| ApiError::bad_request(e.to_string()))?;
-                    let (state, context) = rpc_context.context(&request, module_instance_id).await;
-
-                    let res = (handler)(state, context, request).await;
-
-                    res
-                }))
-                .catch_unwind()
-                .await
-                .map_err(|_| {
-                    error!(
-                        target: LOG_NET_API,
-                        path, "API handler panicked, DO NOT IGNORE, FIX IT!!!"
-                    );
-                    jsonrpsee::core::Error::Call(CallError::Custom(ErrorObject::owned(
-                        500,
-                        "API handler panicked",
-                        None::<()>,
-                    )))
-                })?
-                .map_err(|tokio::time::error::Elapsed { .. }| {
-                    jsonrpsee::core::Error::RequestTimeout
-                })?
-                .map_err(|e| {
-                    jsonrpsee::core::Error::Call(CallError::Custom(ErrorObject::owned(
-                        e.code, e.message, None::<()>,
-                    )))
-                })
-            })
-            .expect("Failed to register async method");
-    }
-}
-
-fn server_endpoints() -> Vec<ApiEndpoint<ConsensusApi>> {
+pub fn server_endpoints() -> Vec<ApiEndpoint<ConsensusApi>> {
     vec![
         api_endpoint! {
             "/transaction",
